% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonnmr_archivecreator.R
\name{moonnmr_archivecreator}
\alias{moonnmr_archivecreator}
\title{Create an archive template}
\usage{
moonnmr_archivecreator()
}
\arguments{
\item{archive}{consists of the columns Date,Name,Projet,Type,Size
See `?golem::get_golem_options` for more details.}
}
\description{
A simple script that creates an empty archive (+archive_backup) template in the current directory
}
